refactor: Restructure program for ETL pipeline
