--- assets/pinyin_tones.py
@@ -1,32 +1,0 @@
-PINYIN_TONES = {
-    "a1": "\u0101",
-    "a2": "\xe1",
-    "a3": "\u01ce",
-    "a4": "\xe0",
-    "a5": "a",
-    "e1": "\u0113",
-    "e2": "\xe9",
-    "e3": "\u011b",
-    "e4": "\xe8",
-    "e5": "e",
-    "i1": "\u012b",
-    "i2": "\xed",
-    "i3": "\u01d0",
-    "i4": "\xec",
-    "i5": "i",
-    "o1": "\u014d",
-    "o2": "\xf3",
-    "o3": "\u01d2",
-    "o4": "\xf2",
-    "o5": "o",
-    "u1": "\u016b",
-    "u2": "\xfa",
-    "u3": "\u01d4",
-    "u4": "\xf9",
-    "u5": "u",
-    "\u00fc1": "\u01d6",
-    "\u00fc2": "\u01d8",
-    "\u00fc3": "\u01da",
-    "\u00fc4": "\u01dc",
-    "\u00fc5": "\u00fc",
-}
