+++ src/resources/assets/junda_frequencies.csv
