+++ src/xiwen/utils/transform.py
@@ -1,17 +1,4 @@
-import os
-import pandas as pd
-from .config import ASSETS_DIR
-
-# Load HSK Hanzi database (unigrams only)
-HSK_HANZI = pd.read_csv(os.path.join(ASSETS_DIR, "hsk30_hanzi.csv"))
-# Replace HSK7-9 with 7 and convert grades to ints for iteration
-HSK_HANZI["HSK Grade"] = HSK_HANZI["HSK Grade"].replace("7-9", 7)
-HSK_HANZI["HSK Grade"] = HSK_HANZI["HSK Grade"].astype(int)
-HSK_SIMP = list(HSK_HANZI["Simplified"])
-HSK_TRAD = list(HSK_HANZI["Traditional"])
-
-
-def partition_hanzi(hanzi_list: list) -> tuple[list]:
+def partition_hanzi(hanzi_list: list, hsk_simp: list, hsk_trad: list) -> tuple[list]:
     """
     Separates hanzi list into sublists based on whether
     they are HSK simplified characters or traditional character equivalents
@@ -21,6 +8,12 @@ def partition_hanzi(hanzi_list: list) -> tuple[list]:
     ----------
     hanzi_list : list
         characters to partition
+
+    hsk_simp : list
+        all simplified characters in HSK
+
+    hsk_trad : list
+        traditional character equivalents to HSK characters
 
     Returns
     -------
@@ -33,8 +26,8 @@ def partition_hanzi(hanzi_list: list) -> tuple[list]:
     outliers : list
         characters not in above lists
     """
-    simp = [zi for zi in hanzi_list if zi in HSK_SIMP]
-    trad = [zi for zi in hanzi_list if zi in HSK_TRAD]
+    simp = [zi for zi in hanzi_list if zi in hsk_simp]
+    trad = [zi for zi in hanzi_list if zi in hsk_trad]
     outliers = [zi for zi in hanzi_list if zi not in simp and zi not in trad]
 
     return simp, trad, outliers
